refactor CategoryType and move it into domain layer
