+++ news/list/src/main/java/com/gradlevv/list/domain/CategoryItem.kt
@@ -2,6 +2,6 @@ package com.gradlevv.list.domain
 
 data class CategoryItem(
     val icon: Int,
-    val type: Int,
+    val type: String,
     val categoryName: Int
 )
